reorganize and small fixes
